+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## 0.26.0 – Apr 24, 2020
-### Featured
 
 ### New
+- companion fields `*_string` for fields that holds unix time values
 
 ### Fixed